Add files via upload
The aim of the project is to build a Machine Learning Model to predict whether an owner will initiate an auto insurance claim in the next year using historical data of those customers. It is a binary classification problem with a huge dataset size. This project is divided into 2 phases, the exploratory data analysis phase and the machine learning modelling phase. It contains extensive markdown and explanation of all the steps taken in the code process in a question answer format.